move ops-masking funcs to ops.masking
